+++ pkg/sysreq/unix.go
@@ -8,6 +8,8 @@ import (
 	"context"
 
 	"golang.org/x/sys/unix"
+
+	"github.com/sourcegraph/sourcegraph/pkg/conf"
 )
 
 func rlimitCheck(ctx context.Context) (problem, fix string, err error) {
@@ -19,7 +21,13 @@ func rlimitCheck(ctx context.Context) (problem, fix string, err error) {
 	}
 
 	if limit.Cur < minLimit {
-		return "Insufficient file descriptor limit", fmt.Sprintf(`Please increase the open file limit by running "ulimit -n %d".`, minLimit), nil
+		fix := fmt.Sprintf(`Please increase the open file limit by running "ulimit -n %d".`, minLimit)
+
+		if conf.IsDeployTypeDockerContainer(conf.DeployType()) {
+			fix = fmt.Sprintf("Add --ulimit nofile=%d:%d to the docker run command", minLimit, minLimit)
+		}
+
+		return "Insufficient file descriptor limit", fix, nil
 	}
 	return
 }